refactor: split tests properties into separate tests
- one test for correctness properties
- one test for aesthetic properties
- improve finding flix files code
- improve comments
