+++ main/test/ca/uwaterloo/flix/tools/fmt/TestFormatterCommon.scala
@@ -1,53 +1,63 @@
-package ca.uwaterloo.flix.api.lsp
+/*
+ * Copyright 2026 Din Jakupi
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package ca.uwaterloo.flix.tools.fmt
 
 import ca.uwaterloo.flix.api.{Flix, Library}
 import ca.uwaterloo.flix.language.CompilationMessage
 import ca.uwaterloo.flix.language.ast.shared.SecurityContext
 import ca.uwaterloo.flix.language.ast.{SyntaxTree, WeededAst}
-import ca.uwaterloo.flix.tools.fmt.PrettyPrinter
 import ca.uwaterloo.flix.util.Formatter.NoFormatter
-import ca.uwaterloo.flix.util.{LibLevel, Options}
-import org.scalatest.Ignore
+import ca.uwaterloo.flix.util.{FileOps, LibLevel, Options}
 import org.scalatest.funsuite.AnyFunSuite
 
 import java.nio.file.{Files, Path, Paths}
-import scala.jdk.CollectionConverters.*
 
 /**
-  * Correctness tests for the Flix code formatter.
+  * Shared fixtures and helpers for the Flix code formatter test suites.
   *
-  * The correctness properties checked here are part of the canonical set
-  * required of a source code formatter.
-  *
+  * The formatter is tested against several properties.
   * Letting
   *   p : S -> C  (parser)
   *   f : C -> S  (formatter)
   *   w : C -> A  (weeder)
   * where `S` is the set of Flix programs accepted by the parser, `C` the set
   * of concrete syntax trees [[SyntaxTree]], and `A` the set of abstract syntax
-  * trees [[WeededAst]], the universal correctness properties are:
+  * trees [[WeededAst]], the properties are:
   *
   *   1. Can format:          forall s in S,  f(p(s)) is defined
   *   2. Idempotency:         forall c in C,  f(p(f(c))) = f(c)
   *   3. Non-destructiveness: forall c in C,  w(c) = w(p(f(c)))
+  *   4. Stability:           forall l in stdlib ++ examples.  f(p(l)) = l
   *
-  * Beyond the correctness properties, the suite also checks one *aesthetic* property.
+  * Properties 1-3 are *correctness* properties tested by [[TestFormatterCorrectness]].
+  * Property 4 is an *aesthetic* property tested by [[TestFormatterStability]].
   *
-  *   4. Stability:   forall l in stdlib ++ examples.  f(p(l)) = l
-  *
-  * The standard library and examples are maintained in a formatted form
-  * and are expected to be the canonical style. This property ensures that the formatter preserves the formatting rules.
-  * Files that are not yet formatted may be listed in [[KnownUnstable]] as exceptions.
+  * This trait provides [[ExampleSamples]], [[StdlibSamples]] and the
+  * parsing ([[reparseAt]]) shared by both tests.
   */
-@Ignore
-class FormatterTest extends AnyFunSuite {
+trait TestFormatterCommon extends AnyFunSuite {
 
   /**
-    * A reparse either yields a SyntaxTree or WeededAst for the given source.
-    * This union type is used to define the `reparse` function in `Sample` as a single function that can be used for both
-    * the SyntaxTree and WeededAst.
+    * The result of reparsing a source both the [[SyntaxTree.Tree]] and the
+    * [[WeededAst.CompilationUnit]] for it.
+    *
+    * A single reparse yields both, so `Sample.reparse` can be one function
+    * regardless of whether a test needs the syntax tree or the weeded unit.
     */
-  private case class Parsed(tree: SyntaxTree.Tree, weeded: WeededAst.CompilationUnit)
+  protected case class Parsed(tree: SyntaxTree.Tree, weeded: WeededAst.CompilationUnit)
 
   /**
     * A sample program for testing
@@ -58,7 +68,7 @@ class FormatterTest extends AnyFunSuite {
     *                after substituting the source for the samples path in the Flix instance and running `check`.
     *                The function is responsible for restoring the Flix instance to its original state after parsing.
     */
-  private case class Sample(
+  protected case class Sample(
     path: String,
     content: String,
     reparse: String => Parsed
@@ -87,10 +97,10 @@ class FormatterTest extends AnyFunSuite {
   }
 
   /**
-    * Every `.flix` file under `examples/`, used as a corpus for all three properties.
+    * Every `.flix` file under `examples/`, used as a corpus for all properties.
     * `apps` and `package-manager` and `datalog/train-schedule.flix` are excluded.
     */
-  private val ExampleSamples: List[Sample] =
+  protected val ExampleSamples: List[Sample] =
     findFlixFiles(
       Paths.get("examples"),
       exclude = Set("apps", "package-manager", "datalog/train-schedule.flix")
@@ -99,27 +109,25 @@ class FormatterTest extends AnyFunSuite {
       Sample(p, content, src => reparseAt(exampleFlix, p, src, restoreTo = None))
     }
 
-  private val StdlibSamples: List[Sample] =
+  /**
+    * Every `.flix` file in the standard library, used as a corpus for all properties.
+    */
+  protected val StdlibSamples: List[Sample] =
     StdlibFiles.map { case (p, content) =>
       Sample(p, content, src => reparseAt(stdlibFlix, p, src, restoreTo = Some(content)))
     }
 
   /**
-    * Returns every `.flix` file under `root`.
-    * Returns the empty list if `root` does not exist.
+    * Find all `.flix` files under `root` excluding the ones that are in the `exclude` set.
+    *
+    * Returns full paths (including `root`) as forward-slash strings, sorted
+    * platform-independently. The `exclude` entries are matched as substrings.
     */
   private def findFlixFiles(root: Path, exclude: Set[String]): List[String] = {
-    if (!Files.exists(root)) Nil
-    else {
-      val stream = Files.walk(root)
-      try stream.iterator().asScala
-        .filter(Files.isRegularFile(_))
-        .map(_.toString.replace("\\", "/"))
-        .filter(_.endsWith(".flix"))
-        .filterNot(p => exclude.exists(p.contains))
-        .toList.sorted
-      finally stream.close()
-    }
+    val files = FileOps.getFlixFilesIn(root, depth = Int.MaxValue)
+    FileOps.sortPlatformIndependently(root, files)
+      .map { case (path, _) => path.normalize().toString() }
+      .filterNot(str => exclude.exists(str.contains))
   }
 
   /**
@@ -131,7 +139,7 @@ class FormatterTest extends AnyFunSuite {
     *
     * Fails the test if `src` does not compile cleanly.
     */
-  private def reparseAt(
+  protected def reparseAt(
     flix: Flix,
     path: String,
     src: String,
@@ -163,101 +171,6 @@ class FormatterTest extends AnyFunSuite {
   }
 
   /**
-    * Property 1 -- Can format: `forall s in S, f(p(s)) is defined`.
-    *
-    * The formatter produces a non-empty string for every [[SyntaxTree]] tree the parser produces.
-    */
-  private def checkCanFormat(samples: List[Sample]): Unit = {
-    for (sample <- samples) {
-      val tree = sample.reparse(sample.content).tree
-      val formatted = PrettyPrinter.format(tree)
-      assert(formatted.nonEmpty, s"Formatter produced empty output for ${sample.path}")
-    }
-  }
-
-  /**
-    * Property 2 -- Idempotency: `forall c in C, f(p(f(c))) = f(c)`.
-    *
-    * Applying the formatter to its own output yields the same output.
-    * The inner `p` accounts for the fact that the formatter produces a string.
-    */
-  private def checkIdempotency(samples: List[Sample]): Unit = {
-    for (sample <- samples) {
-      val tree1 = sample.reparse(sample.content).tree
-      val once = PrettyPrinter.format(tree1)
-      val tree2 = sample.reparse(once).tree
-      val twice = PrettyPrinter.format(tree2)
-      assert(once == twice,
-        s"Formatter is not idempotent for ${sample.path}:\n${firstDivergence(once, twice)}")
-    }
-  }
-
-  /**
-    * Property 3 -- Non-destructiveness: `forall c in C. shape(w(c)) = shape(w(p(f(c))))`.
-    *
-    * The formatter must not change the shape of the [[WeededAst]].
-    * This is  checked by comparing the weeded compilation unit of the original source
-    * against that of the reparsed formatted output, at the kind level only.
-    */
-  private def checkNonDestructive(samples: List[Sample]): Unit = {
-    for (sample <- samples) {
-      val before = sample.reparse(sample.content)
-      val formatted = PrettyPrinter.format(before.tree)
-      val after = sample.reparse(formatted)
-      assert(sameShape(before.weeded, after.weeded),
-        s"Formatter changed the AST shape for ${sample.path}")
-    }
-  }
-
-
-  /**
-    * Property 3 -- Stability: `forall l in stdlib ++ examples. f(p(l)) = l`.
-    *
-    * Aesthetic both the
-    * standard library and the bundled `examples` programs are kept in formatted
-    * form, so formatting any of those files must reproduce its exact source.
-    */
-  private def checkStability(samples: List[Sample]): Unit = {
-    for (sample <- samples) {
-      val formatted = PrettyPrinter.format(sample.reparse(sample.content).tree)
-      val isFixedPoint = formatted == sample.content
-      assert(isFixedPoint,
-        s"Standard library is not preserved by the formatter (f(p(l)) != l) " +
-          s"for ${sample.path}:\n${firstDivergence(sample.content, formatted)}")
-    }
-  }
-
-  // All three properties run on both test corpus, standard library and examples.
-
-  test("PrettyPrinter: can format (examples)") {
-    checkCanFormat(ExampleSamples)
-  }
-  test("PrettyPrinter: can format (stdlib)") {
-    checkCanFormat(StdlibSamples)
-  }
-
-  test("PrettyPrinter: idempotency (examples)") {
-    checkIdempotency(ExampleSamples)
-  }
-  test("PrettyPrinter: idempotency (stdlib)") {
-    checkIdempotency(StdlibSamples)
-  }
-
-  test("PrettyPrinter: non-destructiveness (examples)") {
-    checkNonDestructive(ExampleSamples)
-  }
-  test("PrettyPrinter: non-destructiveness (stdlib)") {
-    checkNonDestructive(StdlibSamples)
-  }
-
-  test("PrettyPrinter: stability (examples)") {
-    checkStability(ExampleSamples)
-  }
-  test("PrettyPrinter: stability (stdlib)") {
-    checkStability(StdlibSamples)
-  }
-
-  /**
     * Finds the syntaxTree for the given URI in the root if it exists.
     *
     * @param root the syntax tree root to search
@@ -282,28 +195,10 @@ class FormatterTest extends AnyFunSuite {
   }
 
   /**
-    * Checks if two [[WeededAst]] have the same shape.
-    * Meaning that they have the same structure of nodes, but not necessarily the same content.
-    * This is used to check the non-destructiveness property of the formatter.
-    *
-    * TODO: This is a very simple check, therefore, find a more robust way to check for the AST integrity.
-    */
-  private def sameShape(a: Any, b: Any): Boolean = (a, b) match {
-    case (x: Iterable[_], y: Iterable[_]) =>
-      // Two collections have the same shape if they have the same length
-      x.size == y.size && x.iterator.zip(y.iterator).forall { case (p, q) => sameShape(p, q) }
-    case (x: Product, y: Product) =>
-      // Two case classes have the same shape if they have the same string prefix.
-      x.productPrefix == y.productPrefix &&
-        x.productIterator.zip(y.productIterator).forall { case (p, q) => sameShape(p, q) }
-    case _ => true
-  }
-
-  /**
     * Finds the first line where two strings diverge and reports surrounding context.
-    * This helps debugging idempotency failures by showing the first point the formatters output differ.
+    * This helps debugging idempotency and stability failures by showing the first point the outputs differ.
     */
-  private def firstDivergence(a: String, b: String): String = {
+  protected def firstDivergence(a: String, b: String): String = {
     val linesA = a.linesIterator.toArray
     val linesB = b.linesIterator.toArray
     val minLen = math.min(linesA.length, linesB.length)
